feat(admin): Implement product management and dashboard enhancements
This commit introduces a comprehensive product management feature for the admin section and enhances the dashboard and navigation.

Key changes:
- **Product Management UI/UX**:
  - Creates a new product list screen (`AdminProductsFragment`) with a professional layout, separating products into "Active" and "Inactive" lists.
  - Adds a category filter dropdown to refine the product list.
  - Implements a product management screen (`ProductManageFragment`) for creating, viewing, and editing product details.
  - Enables activation/deactivation of products directly from the list via a switch.
  - Adds an options menu on each product item for quick edit and delete actions.

- **MVP Refactor for Products**:
  - Refactors the product management feature into a `list` and `manage` package, following a cleaner MVP (Model-View-Presenter) architecture.
  - `AdminProductsPresenter` now handles fetching, filtering, and updating product lists (both active and inactive).
  - `ProductManagePresenter` manages loading product details and handling create, update, and delete operations.

- **Dashboard Enhancements**:
  - The admin dashboard now dynamically displays statistics for total users and products.
  - Refactors `AdminDashboardPresenter` to fetch data from repositories.

- **Navigation and Core Admin Improvements**:
  - Improves `AdminActivity` navigation logic, ensuring the drawer and back stack behave correctly.
  - Implements a logout feature in the settings screen, clearing the user session and navigating to the login screen.
  - Adds HTTP logging interceptor to `RetrofitClient` for easier debugging of network requests.

- **API and Repository Updates**:
  - Adds new API endpoints to fetch inactive products (`/notactive`).
  - `ProductRepository` is updated to fetch both active and inactive products concurrently.
  - `UserRepository` is now a singleton and includes a method to fetch all users.

+++ app/src/main/java/com/example/se1707_prm392_g2_petshop/data/dtos/requests/UpdateProductRequest.java
@@ -4,12 +4,12 @@ public class UpdateProductRequest {
     public int categoryId;
     public String productName;
     public String description ;
-    public int price;
+    public double price;
     public int stock;
     public String imageUrl;
     public boolean isActive;
 
-    public UpdateProductRequest(int categoryId, String productName, String description, int price, int stock, String imageUrl, boolean isActive) {
+    public UpdateProductRequest(int categoryId, String productName, String description, double price, int stock, String imageUrl, boolean isActive) {
         this.categoryId = categoryId;
         this.productName = productName;
         this.description = description;
@@ -43,11 +43,11 @@ public class UpdateProductRequest {
         this.description = description;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
